Re-fixed Gamma Semantic Library by adding isActive to composite state transition firing.
In the past, transition firing operations were surrounded with a choice-else, however, I took that out when the Crossroads model was misbehaving: hierarchicly lower transitions could choose an illegal transition order. The problem was simple to solve, however long it took: by just checking that the composite-state is active, we can ensure the internal transition operations only get executed when appropriate.