#13 Swallow all validator exceptions
so the LSP server does not terminate when an exception is thrown during validation. Usually exceptions are only thrown, when the model is not well-formed, which will either be a syntactical error (will be displayed on the client), or another validation rule will find it. In either case, we can safely ignore all exceptions.